+++ thermostat/devices/inet.py
@@ -1,4 +1,2 @@
 # -*- coding: utf-8 -*-
 """Protocols for remote devices (mainly TCP/IP)."""
-
-# TODO
